minor packages issues
