+++ app/cosi/main.go
@@ -2,14 +2,21 @@ package main
 
 import (
 	"encoding/json"
+	"errors"
 	"flag"
 	"fmt"
+	"golang.org/x/net/context"
 	"io"
+	"io/ioutil"
 	"os"
 	"strings"
 
 	"github.com/dedis/cothority/app"
+	"github.com/dedis/cothority/lib/cosi"
+	"github.com/dedis/cothority/lib/dbg"
+	"github.com/dedis/cothority/lib/network"
 	"github.com/dedis/cothority/lib/sda"
+	"github.com/dedis/crypto/config"
 )
 
 func printUsageAndExit(msg string) {
@@ -89,7 +96,7 @@ func sign(r io.Reader, tomlFileName string) (*sda.CosiResponse, error) {
 	if err != nil {
 		return nil, err
 	}
-	res, err := app.SignStatement(r, el, true)
+	res, err := SignStatement(r, el, true)
 	if err != nil {
 		return nil, err
 	}
@@ -109,4 +116,62 @@ func printSigAsJSON(res *sda.CosiResponse) {
 		fmt.Println("error:", err)
 	}
 	os.Stdout.Write(b)
+}
+
+// SignStatement can be used to sign the contents passed in the io.Reader
+// (pass an io.File or use an strings.NewReader for strings)
+func SignStatement(r io.Reader,
+	el *sda.EntityList,
+	verify bool) (*sda.CosiResponse, error) {
+
+	msgB, err := ioutil.ReadAll(r)
+	if err != nil {
+		return nil, err
+	}
+	// create a throw-away key pair:
+	kp := config.NewKeyPair(network.Suite)
+
+	// create a throw-away entity with an empty  address:
+	e := network.NewEntity(kp.Public, "")
+	client := network.NewSecureTcpHost(kp.Secret, e)
+	req := &sda.CosiRequest{
+		EntityList: el,
+		Message:    msgB,
+	}
+
+	// Connect to the root
+	host := el.List[0]
+	dbg.Lvl3("Opening connection")
+	con, err := client.Open(host)
+	defer client.Close()
+	if err != nil {
+		return nil, err
+	}
+
+	dbg.Lvl3("Sending sign request")
+	// send the request
+	if err := con.Send(context.TODO(), req); err != nil {
+		return nil, err
+	}
+	dbg.Lvl3("Waiting for the response")
+	// wait for the response
+	packet, err := con.Receive(context.TODO())
+	if err != nil {
+		return nil, err
+	}
+	dbg.Lvl3("Recieved response")
+	response, ok := packet.Msg.(sda.CosiResponse)
+	if !ok {
+		return nil, errors.New("Invalid repsonse: Could not cast the " +
+			"received response to the right type")
+	}
+
+	if verify { // verify signature
+		err := cosi.VerifySignature(network.Suite, msgB, el.Aggregate,
+			response.Challenge, response.Response)
+		if err != nil {
+			return nil, err
+		}
+	}
+	return &response, nil
 }
